pyimgui绘制内容拆分
