+++ config/config.py
@@ -91,6 +91,23 @@ class Config(metaclass=SingletonMeta):
             self.log.debug(f"Error accessing JSON data: {e}")
             return None
 
+    def save_json(self, filepath, data):
+        try:
+            with open(filepath, 'w', encoding='utf-8') as file:
+                json.dump(data, file, indent=4, ensure_ascii=False)
+        except (FileNotFoundError, TypeError, json.JSONDecodeError) as e:
+            self.log.debug(f"保存json对象失败: {e}")
+            return None
+
+    # 加载现有数据
+    def load_existing_data(self, filepath):
+        try:
+            with open(filepath, 'r', encoding='utf-8') as file:
+                return json.load(file)
+        except FileNotFoundError as e:
+            self.log.debug(f"读取json对象失败: {e}")
+            return {}
+
     def load_json_as_object(self, file_path, cipher):
         try:
             with open(file_path, 'r', encoding='utf-8') as file:
